Initial management of boxes

+++ config/routes.rb
@@ -6,25 +6,28 @@ Rails.application.routes.draw do
     resources :tenants do
       resources :groups
       resources :users
+      resources :boxes
     end
 
     resources :group_memberships
   end
 
-  namespace :drafts, path: 'drafty' do
-    resources :imports, path: 'importy', only: :create do
-      get :upload_new, path: 'novy', on: :collection
+  namespace :drafts, path: "drafty" do
+    resources :imports, path: "importy", only: :create do
+      get :upload_new, path: "novy", on: :collection
     end
   end
 
-  resources :drafts, path: 'drafty', only: [:index, :show, :destroy] do
+  resources :drafts, path: "drafty", only: %i[index show destroy] do
     post :submit
     post :submit_all, on: :collection
-    delete :destroy_all, path: 'zmazat', on: :collection
+    delete :destroy_all, path: "zmazat", on: :collection
 
     # TODO uncomment later and remove ^ 2 endpoints
     post :submit_all, on: :collection if Rails.env.development?
-    delete :destroy_all, path: 'zmazat', on: :collection if Rails.env.development?
+    if Rails.env.development?
+      delete :destroy_all, path: "zmazat", on: :collection
+    end
   end
 
   resources :sessions do
@@ -32,7 +35,7 @@ Rails.application.routes.draw do
     delete :destroy, on: :collection
   end
 
-  get :auth, path: "prihlasenie", to: 'sessions#login'
+  get :auth, path: "prihlasenie", to: "sessions#login"
   get "auth/google_oauth2/callback", to: "sessions#create"
   get "auth/google_oauth2/failure", to: "sessions#failure"
 
@@ -40,13 +43,11 @@ Rails.application.routes.draw do
 
   class GoodJobAdmin
     def self.matches?(request)
-      admin_ids = ENV.fetch('ADMIN_IDS','').split(',')
+      admin_ids = ENV.fetch("ADMIN_IDS", "").split(",")
 
-      admin_ids.include?(request.session['user_id'].to_s)
+      admin_ids.include?(request.session["user_id"].to_s)
     end
   end
 
-  constraints(GoodJobAdmin) do
-    mount GoodJob::Engine => 'good_job'
-  end
+  constraints(GoodJobAdmin) { mount GoodJob::Engine => "good_job" }
 end
